Reworked Tournament Exceptions

+++ Implementatie/Chessinator/Chessinator.Application/Services/TournamentService.cs
@@ -31,13 +31,13 @@ namespace Chessinator.Application.Services
             // problem here is that the user that is mapped already exists in the database, so
             // we have to get the tracked user from the database and set the tournament user to be the tracked user.
             if (tournamentDto == null)
-                throw new ChessinatorException("tournamentDto is null");
+                throw new InvalidTournamentException("tournamentDto is null");
 
             Tournament tournament = _mapper.Map<Tournament>(tournamentDto);
 
             User trackedUser = await _userRepository.GetUserByIdAsync(tournamentDto.UserId);
             if (trackedUser == null)
-                throw new ChessinatorException("Failed to get user by id.");
+                throw new InvalidTournamentException("Failed to get user by id.");
 
             tournament.User = trackedUser;
             trackedUser.Tournaments.Add(tournament);
@@ -49,7 +49,7 @@ namespace Chessinator.Application.Services
             }
             catch (Exception)
             {
-                throw new ChessinatorException("Failed to create tournament");
+                throw new InvalidTournamentException("Failed to create tournament");
             }
         }
 
@@ -57,7 +57,8 @@ namespace Chessinator.Application.Services
         {
             // tournaments from repository
             List<Tournament> tournaments = await _tournamentRepository.GetTournamentsAsync(userGuid);
-
+            if (tournaments == null)
+                throw new InvalidTournamentException("Failed to get any tournaments");
             // map tournament ENTITY to tournament DTO
             return _mapper.Map<List<TournamentDto>>(tournaments);
         }
@@ -79,17 +80,21 @@ namespace Chessinator.Application.Services
         }
 
         public async Task<bool> DeleteTournamentAsync(Guid tournamentId)
-        { 
+        {
+            if (tournamentId == Guid.Empty)
+            {
+                throw new InvalidTournamentException("Failed to delete tournament");
+            }
             return await _tournamentRepository.DeleteTournamentAsync(tournamentId);
+            
         }
 
         public async Task<TournamentDto> UpdateTournamentAsync(TournamentDto tournamentDto)
         {
             Tournament tournament =_mapper.Map<Tournament>(tournamentDto);
             Tournament updatedTournament = await _tournamentRepository.UpdateTournamentAsync(tournament);
-
             if (updatedTournament == null)
-                throw new ChessinatorException("Failed to update tournament");
+                throw new InvalidTournamentException("Failed to update tournament");
 
             return _mapper.Map<TournamentDto>(updatedTournament);
         }
